Brought phrases linked by a maqqef onto the same line (excluding K/Q)
Used a regex search for:
<w lemma="([^"]+)">(.+)־<\/w>\n\s+<w lemma="([^"]+)">(.+)<\/w>
and replaced with:
<w lemma="$1">$2־<\/w><w lemma="$3">$4<\/w>
twice to bring phrases with a maqqef onto the same line. This eliminates white space between the words in such phrases.
It does not take into account Ketiv/Qere readings. That move will be made once we have a settled way of handling Ketiv/Qere readings in OSIS.